requirements.txt
redis==3.5.3
Flask<2.0.0
gunicorn >= 19.9.0, < 20.2.0
pyyaml>=5.3,<5.5
protobuf>=3.7.0,<4
requests>=2.17.3,<3.0.0
numpy>=1.19.0,<2.0.0
dill>=0.3.4
seldon-core==1.11.2
scikit-learn==1.0.1
